+++ CHANGELOG.md
@@ -19,7 +19,9 @@ verbose output
 ### Fixed
 - Cascade (container-only) start issue with no private registry
 - Non-shipyard docker image node prep with new azure-storage package
-- Temp file creation used for environment variables now works cross-platform
+- Cross-platform fixes:
+  - Temp file creation used for environment variables
+  - SSH tunnel creation disabled on Windows if public key is not supplied
 
 ### Removed
 - `gpu`:`nvidia_driver`:`version` property removed from pool configuration